chore(core): ensure defaults for internal settings
